requirements.txt
nose
biopython